requirements.txt
Flask
gunicorn
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-api-core
requests
python-dotenv
